+++ src/psd.ts
@@ -263,14 +263,14 @@ export interface Psd extends LayerAdditionalInfo {
 export interface ReadOptions {
 	/** does not load layer image data */
 	skipLayerImageData?: boolean;
-	/** does not load layer mask data */
-	skipLayerMaskData?: boolean;
-	/** don't treat "channel not supported" as a fatal error */
-	channelNotSupportedNotFatal?: boolean;
 	/** does not load composite image data */
 	skipCompositeImageData?: boolean;
 	/** does not load thumbnail */
 	skipThumbnail?: boolean;
+	/** throws exception if features are missing */
+	throwForMissingFeatures?: boolean;
+	/** logs if features are missing */
+	logMissingFeatures?: boolean;
 }
 
 export interface WriteOptions {